NickAkhmetov/CAT-937 Sort publications by their publication date, convert related hooks to TypeScript
