+++ CHANGELOG-cat-937.md
@@ -1,0 +1,1 @@
+- Sort publications by their date of publication.
